Extract getAllResponse funnctionality to the utils folder
